+++ lib/features/notes/notes.dart
@@ -1,1 +1,0 @@
-export 'package:note_taking_app/features/notes/presentation/screens/note_editor_screen.dart';