+++ lib/features/settings/settings.dart
@@ -1,2 +1,0 @@
-export 'package:note_taking_app/features/settings/presentation/screens/settings_screen.dart';
-export '../../screens/app_lock_screen.dart';